Update cveda_psytools_derive.R
Added an iteration selection procedure after derivation - As in Imagen this takes the first complete attempt for tasks and the last complete attempt for everything else. In a special case for cVeda the 3 different age band slots for a PSC are collapsed and the iteration recreated from the completed timestamp.